Move processing into own component.
