+++ src/xcode/ENA/ENA/Source/Services/TicketValidation/TicketValidating.swift
@@ -5,12 +5,7 @@
 import Foundation
 
 protocol TicketValidating {
-
-	init(
-		with initializationData: TicketValidationInitializationData,
-		restServiceProvider: RestServiceProviding
-	)
-
+	
 	var initializationData: TicketValidationInitializationData { get }
 
 	func initialize(
