+++ src/xcode/ENA/ENA/Source/Services/TicketValidation/TicketValidation.swift
@@ -5,16 +5,32 @@
 import Foundation
 import ENASecurity
 
+struct ServiceIdentityRequestResult {
+	let validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC: [JSONWebKey]
+	let validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM: [JSONWebKey]
+	let validationServiceSignKeyJwkSet: [JSONWebKey]
+}
+
+enum ServiceIdentityRequestError: Error {
+	case VS_ID_NO_ENC_KEY
+	case VS_ID_NO_SIGN_KEY
+	case VS_ID_EMPTY_X5C
+	case REST_SERVICE_ERROR(ServiceError<ServiceIdentityDocumentResourceError>)
+	case UNKOWN
+}
+
 final class TicketValidation: TicketValidating {
 	
 	// MARK: - Protocol TicketValidating
 
 	init(
 		with initializationData: TicketValidationInitializationData,
-		restServiceProvider: RestServiceProviding
+		restServiceProvider: RestServiceProviding,
+		serviceIdentityProcessor: ServiceIdentityDocumentProcessing
 	) {
 		self.initializationData = initializationData
 		self.restServiceProvider = restServiceProvider
+		self.serviceIdentityProcessor = serviceIdentityProcessor
 	}
 
 	var initializationData: TicketValidationInitializationData
@@ -46,27 +62,12 @@ final class TicketValidation: TicketValidating {
 	func cancel() {
 
 	}
-
-	struct ServiceIdentityRequestResult {
-		let validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC: [JSONWebKey]
-		let validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM: [JSONWebKey]
-		let validationServiceSignKeyJwkSet: [JSONWebKey]
-	}
-	
-	enum ServiceIdentityRequestError: Error {
-		case VS_ID_NO_ENC_KEY
-		case VS_ID_NO_SIGN_KEY
-		case VS_ID_EMPTY_X5C
-		case REST_SERVICE_ERROR(ServiceError<ServiceIdentityDocumentResourceError>)
-		case UNKOWN
-	}
 	
 	func requestServiceIdentityDocument(
 		validationServiceData: ValidationServiceData,
 		validationServiceJwkSet: [JSONWebKey],
 		completion: @escaping (Result<ServiceIdentityRequestResult, ServiceIdentityRequestError>) -> Void
 	) {
-		
 		guard let url = URL(string: validationServiceData.serviceEndpoint) else {
 			completion(.failure(.UNKOWN))
 			return
@@ -74,75 +75,14 @@ final class TicketValidation: TicketValidating {
 		
 		let resource = ServiceIdentityDocumentResource(endpointUrl: url)
 		
-		restServiceProvider.load(resource) { result in
+		restServiceProvider.load(resource) { [weak self] result in
 			switch result {
 			case .success(let serviceIdentityDocument):
-                
-				// 2. Verifiy JWKs
-				for verificationMethod in serviceIdentityDocument.verificationMethod {
-                    if let publicKeyJwk = verificationMethod.publicKeyJwk, publicKeyJwk.x5c.isEmpty {
-						Log.error("Verify JWKs failed", log: .ticketValidation)
-						completion(.failure(.VS_ID_EMPTY_X5C))
-					}
-				}
-                
-                // 3. Find verificationMethodsForRSAOAEPWithSHA256AESCBC
-                let cbcRegEx = "ValidationServiceEncScheme-RSAOAEPWithSHA256AESCBC$"
-                let verificationMethodsForRSAOAEPWithSHA256AESCBC = serviceIdentityDocument.verificationMethod.first { verificationMethod in
-                    let regExExists = verificationMethod.id.check(regex: cbcRegEx)
-                    return regExExists && verificationMethod.verificationMethods != nil
-                }?.verificationMethods ?? []
-				
-				// 4. Find validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC
-				let validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC = serviceIdentityDocument.verificationMethod.filter { verificationMethod in
-					verificationMethodsForRSAOAEPWithSHA256AESCBC.contains(verificationMethod.id)
-				}.compactMap { verificationMethod in
-					verificationMethod.publicKeyJwk
-				}
-				Log.debug("Found validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC: \(private: validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC)", log: .ticketValidation)
-				
-				// 5. Find verificationMethodsForRSAOAEPWithSHA256AESGCM
-				let gcmRegEx = "ValidationServiceEncScheme-RSAOAEPWithSHA256AESGCM$"
-				let verificationMethodsForRSAOAEPWithSHA256AESGCM = serviceIdentityDocument.verificationMethod.first { verificationMethod in
-					let regExExists = verificationMethod.id.check(regex: gcmRegEx)
-					return regExExists && verificationMethod.verificationMethods != nil
-				}?.verificationMethods ?? []
-				
-				// 6. Find validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM
-				let validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM = serviceIdentityDocument.verificationMethod.filter { verificationMethod in
-					verificationMethodsForRSAOAEPWithSHA256AESGCM.contains(verificationMethod.id)
-				}.compactMap { verificationMethod in
-					verificationMethod.publicKeyJwk
-				}
-				Log.debug("Found validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM: \(private: validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM)")
-				
-				// 7. Check encryption keys
-				if validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC.isEmpty && validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM.isEmpty {
-					completion(.failure(.VS_ID_NO_ENC_KEY))
-					return
-				}
-				
-				// 8. Find validationServiceSignKeyJwkSet
-				let regex = "ValidationServiceSignKey-\\d+$"
-				let validationServiceSignKeyJwkSet = serviceIdentityDocument.verificationMethod.filter { verificationMethod in
-					verificationMethod.id.check(regex: regex)
-				}.compactMap { verificationMethod in
-					verificationMethod.publicKeyJwk
-				}
-				
-				if validationServiceSignKeyJwkSet.isEmpty {
-					completion(.failure(.VS_ID_NO_SIGN_KEY))
-					return
-				}
-				Log.debug("Found validationServiceSignKeyJwkSet: \(private: validationServiceSignKeyJwkSet)")
-
-				let result = ServiceIdentityRequestResult(
-					validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC: validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC,
-					validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM: validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM,
-					validationServiceSignKeyJwkSet: validationServiceSignKeyJwkSet
+				self?.serviceIdentityProcessor.process(
+					validationServiceJwkSet: validationServiceJwkSet,
+					serviceIdentityDocument: serviceIdentityDocument,
+					completion: completion
 				)
-				completion(.success(result))
-				
 			case .failure(let error):
 				completion(.failure(.REST_SERVICE_ERROR(error)))
 			}
@@ -152,5 +92,6 @@ final class TicketValidation: TicketValidating {
 	// MARK: - Private
 
 	private let restServiceProvider: RestServiceProviding
+	private let serviceIdentityProcessor: ServiceIdentityDocumentProcessing
 	
 }
